Refactoring to /app folder, updated deps
