+++ app/main.py
@@ -1,12 +1,13 @@
-import os
-import discord
-from gitclient import GitClient
-from dotenv import load_dotenv
-import logging
-from typing import Union
-from fastapi import FastAPI
 import asyncio
-from pydantic import BaseModel
+import logging
+import os
+
+import discord
+from dotenv import load_dotenv
+from fastapi import FastAPI
+
+from app.client.gitclient import GitClient
+from app.schemas.updates import Updates
 
 app = FastAPI()
 load_dotenv()
@@ -21,24 +22,21 @@ intents.message_content = True
 
 client = GitClient(intents=intents)
 
+
 @app.on_event("startup")
 async def startup_event():
     logging.info("Starting up")
     asyncio.create_task(client.start(token))
 
+
 @app.get("/")
 async def read_root():
+    logging.info("Received request on /")
     return {"Hello": "World"}
 
-class Updates(BaseModel):
-    EventType: str
-    EventDescription: str 
-    EventUrl : str
-    EventReceipients : int
 
 @app.post("/eventupdates")
 async def post_event_updates(updates: Updates):
+    logging.info(f"Received request on /eventupdates: {updates}")
     await client.on_update_event(updates.EventType, updates.EventDescription, updates.EventUrl, updates.EventReceipients)
     return updates
-
-
