+++ app/schemas/updates.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class Updates(BaseModel):
+    EventType: str
+    EventDescription: str 
+    EventUrl : str
+    EventReceipients : int
