+++ requirements.txt
@@ -2,88 +2,24 @@ aiohttp==3.8.5
 aiosignal==1.3.1
 annotated-types==0.5.0
 anyio==3.7.1
-asttokens==2.2.1
 async-timeout==4.0.3
 attrs==23.1.0
-backcall==0.2.0
-certifi==2022.6.15.1
 charset-normalizer==3.3.0
-click==8.1.3
-click-plugins==1.1.1
-cligj==0.7.2
-colorama==0.4.5
-comm==0.1.3
-debugpy==1.6.7
-decorator==5.1.1
+click==8.1.7
+colorama==0.4.6
 discord.py==2.3.2
-docopt==0.6.2
-et-xmlfile==1.1.0
 exceptiongroup==1.1.3
-executing==1.2.0
 fastapi==0.103.2
-Fiona==1.8.22
-Flask==2.2.2
-Flask-Cors==3.0.10
-Flask-MySQLdb==1.0.1
 frozenlist==1.4.0
-geopandas==0.12.0
 h11==0.14.0
-httptools==0.6.0
 idna==3.4
-importlib-metadata==5.0.0
-ipykernel==6.22.0
-ipython==8.12.0
-itsdangerous==2.1.2
-jedi==0.18.2
-Jinja2==3.1.2
-jupyter_client==8.1.0
-jupyter_core==5.3.0
-MarkupSafe==2.1.1
-matplotlib-inline==0.1.6
 multidict==6.0.4
-munch==2.5.0
-mysqlclient==2.1.1
-nest-asyncio==1.5.6
-numpy==1.23.4
-opencv-python==4.6.0.66
-openpyxl==3.0.10
-packaging==21.3
-pandas==1.5.1
-parso==0.8.3
-pickleshare==0.7.5
-pipreqs==0.4.11
-platformdirs==3.2.0
-prompt-toolkit==3.0.38
-psutil==5.9.4
-pure-eval==0.2.2
+packaging==23.2
 pydantic==2.4.2
 pydantic_core==2.10.1
-Pygments==2.15.0
-pyparsing==3.0.9
-pyproj==3.4.0
-pyshp==2.3.1
-python-dateutil==2.8.2
 python-dotenv==1.0.0
-pytz==2022.2.1
-pywin32==306
-PyYAML==6.0.1
-pyzmq==25.0.2
-requests==2.28.1
-Shapely==1.8.5.post1
-six==1.16.0
 sniffio==1.3.0
-stack-data==0.6.2
 starlette==0.27.0
-tornado==6.2
-traitlets==5.9.0
 typing_extensions==4.8.0
-urllib3==1.26.12
 uvicorn==0.23.2
-watchfiles==0.20.0
-wcwidth==0.2.6
-websockets==11.0.3
-Werkzeug==2.2.2
-XlsxWriter==3.0.3
-yarg==0.1.9
 yarl==1.9.2
-zipp==3.9.0
